requirements.txt
flask
langchain
flask-wtf
flask-bootstrap
pypdf
langchain
chromadb # Vector storage
pytest
boto3
langchain-community
langchain_ollama
langchain_chroma
filetype
pypdf2
pytesseract
fpdf
pillow
pdf2image
poppler-utils
transformers
torch
logging